refactor: splict `storage_manager` into several files
