+++ src/storage/storage_manager/mod.rs
@@ -1,0 +1,8 @@
+//! The storage manager.
+
+mod manager;
+
+pub use manager::StorageManager;
+
+#[cfg(test)]
+mod test;
